requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-braces==1.8.1
django-floppyforms==1.5.2
gunicorn==19.3.0
psycopg2==2.6.1
six==1.9.0
static3==0.6.1
wheel==0.24.0
